+++ src/collective/exportimport/tests/test_import.py
@@ -11,6 +11,8 @@ from plone.app.testing import SITE_OWNER_NAME
 from plone.app.testing import SITE_OWNER_PASSWORD
 from plone.namedfile.file import NamedImage
 from plone.namedfile.file import NamedBlobImage
+from Products.CMFPlone.interfaces.constrains import ENABLED
+from Products.CMFPlone.interfaces.constrains import ISelectableConstrainTypes
 from Products.CMFPlone.tests import dummy
 
 import json
@@ -206,23 +208,163 @@ class TestImport(unittest.TestCase):
         self.assertEqual(new_doc.portal_type, "Document")
         self.assertEqual(new_doc.UID(), doc_uid)
 
-        # See what happens when we import it a second time.
+        # When we import it a second time is is ignored by default.
         original_ids = portal.contentIds()
         browser = self.open_page("@@import_content")
         upload = browser.getControl(name="jsonfile")
         upload.add_file(raw_data, "application/json", "Document.json")
+        browser.getForm(action="@@import_content").submit()
+        self.assertIn("Imported 0 items in 0 seconds", browser.contents)
+
+        # Now we ignore it
+        original_ids = portal.contentIds()
+        browser = self.open_page("@@import_content")
+        upload = browser.getControl(name="jsonfile")
+        upload.add_file(raw_data, "application/json", "Document.json")
+        browser.getControl(name="handle_existing_content").value = ["3"]
         browser.getForm(action="@@import_content").submit()
         self.assertIn("Imported 1 items", browser.contents)
 
         # A second document should be there.
         new_ids = [docid for docid in portal.contentIds() if docid not in original_ids]
         self.assertEqual(len(new_ids), 1)
+        self.assertEqual(len(portal.contentIds()), 2)
         doc2 = portal[new_ids[0]]
         # doc2 is the same as the original
         self.assertEqual(doc2.Title(), "Document 1")
         self.assertEqual(doc2.portal_type, "Document")
         # except for the UID
         self.assertNotEqual(doc2.UID(), doc_uid)
+
+    def test_import_content_update(self):
+        # First create some content.
+        app = self.layer["app"]
+        portal = self.layer["portal"]
+        login(app, SITE_OWNER_NAME)
+        doc = api.content.create(
+            container=portal, type="Document", id="doc1", title="Document 1"
+        )
+        doc_uid = doc.UID()
+        transaction.commit()
+
+        # Now export it.
+        browser = self.open_page("@@export_content")
+        browser.getControl(name="portal_type").value = ["Document"]
+        browser.getForm(action="@@export_content").submit(name="submit")
+
+        # We should have gotten json.
+        raw_data = browser.contents
+        if not browser.contents:
+            raw_data = DATA[-1]
+
+        # Remove the added content.
+        api.content.delete(doc)
+        transaction.commit()
+        self.assertNotIn("doc1", portal.contentIds())
+
+        # Now import it.
+        browser = self.open_page("@@import_content")
+        upload = browser.getControl(name="jsonfile")
+        upload.add_file(raw_data, "application/json", "Document.json")
+        browser.getForm(action="@@import_content").submit()
+        self.assertIn("Imported 1 items", browser.contents)
+
+        # The document should be back.
+        self.assertIn("doc1", portal.contentIds())
+        new_doc = portal["doc1"]
+        self.assertEqual(new_doc.Title(), "Document 1")
+        self.assertEqual(new_doc.portal_type, "Document")
+        self.assertEqual(new_doc.UID(), doc_uid)
+
+        # When we import and update it with some changed data.
+        data = json.loads(raw_data)
+        data[0]["title"] = "A different title"
+        data[0].pop("description")
+        changed_raw_data = json.dumps(data)
+
+        browser = self.open_page("@@import_content")
+        upload = browser.getControl(name="jsonfile")
+        upload.add_file(changed_raw_data.encode(), "application/json", "Document.json")
+        browser.getControl(name="handle_existing_content").value = ["2"]  # update!
+        browser.getForm(action="@@import_content").submit()
+        self.assertIn("Imported 1 items in 0 seconds", browser.contents)
+
+        # new_doc now has a updated title
+        new_doc = portal["doc1"]
+        self.assertEqual(len(portal.contentIds()), 1)
+        self.assertEqual(new_doc.Title(), "A different title")
+        self.assertEqual(new_doc.portal_type, "Document")
+        # The UID is still the same
+        self.assertEqual(new_doc.UID(), doc_uid)
+
+    def test_import_content_replace(self):
+        # First create some content.
+        app = self.layer["app"]
+        portal = self.layer["portal"]
+        login(app, SITE_OWNER_NAME)
+        doc = api.content.create(
+            container=portal,
+            type="Document",
+            id="doc1",
+            title="Document 1",
+            description="A Description",
+        )
+        doc_uid = doc.UID()
+        transaction.commit()
+
+        # Now export it.
+        browser = self.open_page("@@export_content")
+        browser.getControl(name="portal_type").value = ["Document"]
+        browser.getForm(action="@@export_content").submit(name="submit")
+
+        # We should have gotten json.
+        raw_data = browser.contents
+        if not browser.contents:
+            raw_data = DATA[-1]
+
+        # Remove the added content.
+        api.content.delete(doc)
+        transaction.commit()
+        self.assertNotIn("doc1", portal.contentIds())
+
+        # Now import it.
+        browser = self.open_page("@@import_content")
+        upload = browser.getControl(name="jsonfile")
+        upload.add_file(raw_data, "application/json", "Document.json")
+        browser.getForm(action="@@import_content").submit()
+        self.assertIn("Imported 1 items", browser.contents)
+
+        # The document should be back.
+        self.assertIn("doc1", portal.contentIds())
+        new_doc = portal["doc1"]
+        self.assertEqual(new_doc.Title(), "Document 1")
+        self.assertEqual(new_doc.Description(), "A Description")
+        self.assertEqual(new_doc.portal_type, "Document")
+        self.assertEqual(new_doc.UID(), doc_uid)
+
+        # When we import and replace it with different data.
+        data = json.loads(raw_data)
+        data[0]["title"] = "A different title"
+        data[0].pop("description")
+        changed_raw_data = json.dumps(data)
+
+        browser = self.open_page("@@import_content")
+        upload = browser.getControl(name="jsonfile")
+        upload.add_file(changed_raw_data.encode(), "application/json", "Document.json")
+        browser.getControl(name="handle_existing_content").value = ["1"]  # replace!
+        browser.getForm(action="@@import_content").submit()
+        self.assertIn("Imported 1 items in 0 seconds", browser.contents)
+
+        # still only one item
+        self.assertEqual(len(portal.contentIds()), 1)
+        # new_doc now has a updated title
+        new_doc = portal["doc1"]
+        self.assertEqual(new_doc.Title(), "A different title")
+        # description is new (=empty)
+        self.assertEqual(new_doc.Description(), "")
+        self.assertEqual(new_doc.portal_type, "Document")
+        # The UID is still the same
+        self.assertEqual(new_doc.UID(), doc_uid)
 
     def test_import_content_with_missing_folder(self):
         # First create some content.
@@ -512,6 +654,82 @@ class TestImport(unittest.TestCase):
         self.assertEqual(portal["events"].portal_type, "Folder")
         self.assertEqual(portal["image"].image.data, dummy.Image().data)
 
+    def test_import_imports_but_ignores_constrains(self):
+        """Constrains are exported and imported but not checked during import
+        """
+        # First create some content to export.
+        app = self.layer["app"]
+        portal = self.layer["portal"]
+        login(app, SITE_OWNER_NAME)
+        self.create_demo_content()
+        self.assertIn("events", portal.contentIds())
+        self.assertIn("image", portal.contentIds())
+
+        # create a collection in self.about
+        collection = api.content.create(
+            container=self.about,
+            type="Collection",
+            id="collection",
+            title=u"Collection",
+        )
+        # constrain self.about to only allow documents
+        constrains = ISelectableConstrainTypes(self.about)
+        constrains.setConstrainTypesMode(ENABLED)
+        constrains.setLocallyAllowedTypes(["Document"])
+        constrains.setImmediatelyAddableTypes(["Document"])
+        from plone.api.exc import InvalidParameterError
+        with self.assertRaises(InvalidParameterError):
+            api.content.create(
+                container=self.about,
+                type="Collection",
+                id="collection2",
+                title=u"Collection 2",
+            )
+        transaction.commit()
+
+        # Now export the complete portal.
+        browser = self.open_page("@@export_content")
+        browser.getControl(name="portal_type").value = [
+            "Folder",
+            "Image",
+            "Link",
+            "Document",
+            "Collection",
+        ]
+        browser.getForm(action="@@export_content").submit(name="submit")
+        contents = browser.contents
+        if not browser.contents:
+            contents = DATA[-1]
+
+        data = json.loads(contents)
+        self.assertEqual(len(data), 7)
+
+        # Remove the added content.
+        self.remove_demo_content()
+        transaction.commit()
+        self.assertNotIn("events", portal.contentIds())
+
+        # Now import it.
+        browser = self.open_page("@@import_content")
+        upload = browser.getControl(name="jsonfile")
+        upload.add_file(contents, "application/json", "Document.json")
+        browser.getForm(action="@@import_content").submit()
+        self.assertIn("Imported 7 items", browser.contents)
+
+        # The collection is imported despite the constrain
+        constrains = ISelectableConstrainTypes(portal["about"])
+        self.assertEqual(constrains.getConstrainTypesMode(), ENABLED)
+        self.assertEqual(constrains.getLocallyAllowedTypes(), ["Document"])
+        self.assertEqual(constrains.getLocallyAllowedTypes(), ["Document"])
+        self.assertEqual(portal["about"]["collection"].portal_type, "Collection")
+        with self.assertRaises(InvalidParameterError):
+            api.content.create(
+                container=portal["about"],
+                type="Collection",
+                id="collection2",
+                title=u"Collection 2",
+            )
+
     def _disabled_test_import_blob_path(self):
         # This test is disabled, because the demo storage
         # has no 'fshelper' from which we can ask the blob path.
@@ -708,3 +926,67 @@ class TestImport(unittest.TestCase):
         self.assertEqual(ordered.getObjectPosition("doc1"), 0)
         self.assertEqual(ordered.getObjectPosition("doc2"), 1)
         self.assertEqual(ordered.getObjectPosition("doc3"), 2)
+
+    def test_import_localroles(self):
+        # First create some content.
+        app = self.layer["app"]
+        portal = self.layer["portal"]
+        login(app, SITE_OWNER_NAME)
+        folder = api.content.create(
+            container=portal, type="Folder", id="folder", title="Folder"
+        )
+        folder.__ac_local_roles_block__ = 1
+        doc1 = api.content.create(
+            container=portal, type="Document", id="doc1", title="Document"
+        )
+        doc2 = api.content.create(
+            container=folder, type="Document", id="doc2", title="Document"
+        )
+        api.user.create(
+            username="peter",
+            email="peter@example.org",
+            password="secret",
+            roles=("Member",),
+        )
+        api.user.grant_roles(username="peter", obj=doc1, roles=["Reviewer"])
+        api.user.grant_roles(username="peter", obj=doc2, roles=["Owner"])
+        transaction.commit()
+
+        # Export.
+        browser = self.open_page("@@export_localroles")
+        browser.getForm(action="@@export_localroles").submit(name="form.submitted")
+        raw_data = browser.contents
+        if not browser.contents:
+            raw_data = DATA[-1]
+
+        data = json.loads(raw_data)
+        self.assertEqual(data[0]["block"], 1)
+        self.assertEqual(data[0]["localroles"], {"admin": ["Owner"]})
+        self.assertEqual(data[1]["localroles"], {"admin": ["Owner"], "peter": ["Owner"]})
+        self.assertEqual(data[2]["localroles"], {"admin": ["Owner"], "peter": ["Reviewer"]})
+        self.assertNotIn("block", data[1])
+        self.assertNotIn("block", data[2])
+
+        # remove local roles
+        del folder.__ac_local_roles_block__
+        api.user.revoke_roles(username="peter", obj=doc1, roles=["Reviewer"])
+        api.user.revoke_roles(username="peter", obj=doc2, roles=["Owner"])
+        self.assertEqual(doc1.__ac_local_roles__, {'admin': ['Owner']})
+        folder.reindexObjectSecurity()
+        doc1.reindexObjectSecurity()
+        doc2.reindexObjectSecurity()
+
+        # Import and check.
+        browser = self.open_page("@@import_localroles")
+        upload = browser.getControl(name="jsonfile")
+        upload.add_file(raw_data, "application/json", "localroles.json")
+        browser.getForm(action="@@import_localroles").submit()
+        self.assertIn("Imported 3 localroles", browser.contents)
+        # The documents have the original local roles again.
+        self.assertEqual(folder.__ac_local_roles_block__, 1)
+        self.assertEqual(doc1.__ac_local_roles__, {'admin': ['Owner'], 'peter': ['Reviewer']})
+        # permissions are reindexed
+        self.assertTrue(api.user.has_permission("Review portal content", username="peter", obj=doc1))
+        self.assertFalse(api.user.has_permission("Modify portal content", username="peter", obj=doc1))
+        self.assertFalse(api.user.has_permission("Review portal content", username="peter", obj=doc2))
+        self.assertTrue(api.user.has_permission("Modify portal content", username="peter", obj=doc2))